new admin page drafted
