+++ src/styles/Admin.css
@@ -1,0 +1,3 @@
+#red {
+    color: red;
+}
